Remove reliance on ERBTracker from rails
Use dependency tracker that does not rely on a rails dependency tracker.
This prevents problems if rails chooses to remove or change ERBTracker.
